requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.30.0

# Object Detection
torchmetrics>=1.0.0
pycocotools>=2.0.6

# Data Processing
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.8.0

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Configuration
PyYAML>=6.0

# Development
tqdm>=4.65.0
scikit-learn>=1.3.0

# Optional (for specific models)
# ultralytics>=8.0.0  # For YOLO

